### 2020-09-01 #2609 @mikldk
@VincentAlcazer Thank you for this submission. Do you have any reviewers in mind (e.g. from [the list of people](https://bit.ly/joss-reviewers) mentioned above)?

The title of the paper has "a R package". This should be changed to "an R package" (or maybe even "An R package").
