### 2020-09-01 #2609 @VincentAlcazer
@mikldk Thank you for your comment. Based on the list, I could suggest @adithirgis who seems to have an expertise in data analysis and Shiny apps, and @nistara who seems to be comfortable with health-related areas.

I am making the change for the title.
