+++ README.Rmd
@@ -19,7 +19,7 @@ knitr::opts_chunk$set(
 [![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
 <!-- badges: end -->
 
-`StatAid` is a free open-source software provided as a R package allowing clinicians and researchers to perform statistic analysis through an intuitive graphical interface. It has been developed with the R software, using the [Shiny package](https://shiny.rstudio.com/). [Golem](https://github.com/ThinkR-open/golem) has been used for package compilation and deployment.  
+`StatAid` is a free open-source software provided as an R package allowing clinicians and researchers to perform statistic analysis through an intuitive graphical interface. It has been developed with the R software, using the [Shiny package](https://shiny.rstudio.com/). [Golem](https://github.com/ThinkR-open/golem) has been used for package compilation and deployment.  
 
  The software guide the users through the steps of a good data analysis, including multiple features:
 <ul><li> Exploratory data analysis: distribution, count, missing-values and outliers check  </li>
