### 2020-09-01 #2609 @adithirgis
Hi,

Thanks, I would be interested to review this. I will be able to start only
in a couple of days though. Is that ok?

Thanks!

On Wed 2 Sep, 2020, 1:20 AM Mikkel Meyer Andersen, <notifications@github.com>
wrote:

> @adithirgis <https://github.com/adithirgis>: Would you be interested in
> reviewing this submission to The Journal of Open Source Software
> <https://joss.theoj.org/>? Reviews are open and based on a checklist. The
> reviewer guidelines are available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. If you
> have any questions or concerns please let me know.
>
