### 2020-09-02 #2630 @whedon
[REVIEW]: StatAid: a R package with a graphical user interface for data analysis
**Submitting author:** @VincentAlcazer (<a href="http://orcid.org/0000-0003-1843-6286">Vincent Alcazer</a>)
**Repository:** <a href="https://github.com/VincentAlcazer/StatAid" target ="_blank">https://github.com/VincentAlcazer/StatAid</a>
**Version:** v1.1.2
