### 2020-09-07 VincentAlcazer/StatAid#1 @adithirgis
JOSS Review
Hello, Thanks for submitting this to JOSS and inviting me to review. I found the software very useful. Very good work. Congratulations.  I am adding a few comments for this below:

1.  I am not able to see the page correctly there is a overlap between the mobile/drag and drop menu in the right and the table area for the Data Loading tab. (I am using a Windows system). Any thoughts about this? Can the table to be placed a little lower?
![image](https://user-images.githubusercontent.com/63432077/92202807-63cee100-ee9d-11ea-9129-e57637828607.png)
2.  I tried to run this app without the any data upload : It shows me this error for all combinations dropdown items selected and running the _Descriptive Table_ analysis in Data Exploration tab. Am I selecting the wrong options?
![image](https://user-images.githubusercontent.com/63432077/92203283-67169c80-ee9e-11ea-90e2-b145328ea026.png)
3. Let me know if I am doing anything wrong here: I am trying to run GAM and LOESS for two numeric variables I get this error.
![image](https://user-images.githubusercontent.com/63432077/92205451-53ba0000-eea3-11ea-93b0-a9af401af7e6.png)
![image](https://user-images.githubusercontent.com/63432077/92205631-bb704b00-eea3-11ea-9596-7c4f14a59057.png)

4. Please add Community Guidelines to the [README.md](https://github.com/VincentAlcazer/StatAid/blob/master/README.md) file: stating how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
5. I see that the description and walk-through is well written inside the application interface. It would be very helpful for other users not in the medical/clinical research to use this package for their own research. A descriptive README.md containing example usage and functionality documentation would be very helpful. I strongly suggest to write a descriptive [README.md](https://github.com/VincentAlcazer/StatAid/blob/master/README.md) in your repository. Also, this might be a general comment: Is there a possibility to add description of the pre-loaded data base columns to understand the working of the package better.
6. Are there any other software similar to StatAid existing already? It will be useful to add a comparison here.
7. You can also list the academic papers in the API Documentation/ [README.md](https://github.com/VincentAlcazer/StatAid/blob/master/README.md) where StatAid has been cited. As stated in the [paper.md](https://github.com/VincentAlcazer/StatAid/blob/master/paper.md):  _"It has already been used by multiple researchers and students for their PhD or medical doctors for their thesis."_
8.  I think the good practice of writing cleaner consistent code can be implemented here using these wonderful resources (I too found out during my paper review).  Implementing 80 or 120 characters in a line and spacing consistency, see http://adv-r.had.co.nz/Style.html and https://devguide.ropensci.org/building.html#code-style.
9. I would also suggest to do a check using this command- `devtools::check()`. I came across three warnings(). To solve those you can add some of the build ignore files into this [file](https://github.com/VincentAlcazer/StatAid/blob/master/.Rbuildignore). There is also a import from namespaces issue/warning which can be probably solved using [this.](https://stackoverflow.com/questions/58651058/roxygen-2-do-not-edit-by-hand-vs-notes-add-to-namespaces)
10. A minor addition in the installation [here](https://github.com/VincentAlcazer/StatAid) would be more clear:
````
# StatAid::run_app()
````
11. I would recommend to more software usage in the [paper.md file](https://github.com/VincentAlcazer/StatAid/blob/master/paper.md).
12. I would also suggest to add a few more automated tests other than the recommended one in `golem`. Please let me know if they are already added and if I skipped it.
13. I downloaded the data sets and found that the IRIS data set was not giving desired results I was not able to select any numeric column for regression and correlation or run lm/gam/loess. Are there any other checks to be done before we enter new data into the software other than those mentioned in the software Introduction tab.
![image](https://user-images.githubusercontent.com/63432077/92217056-caabc480-eeb4-11ea-8561-a1a9311ed746.png)

Let me know if you have any further questions.
Thanks again,
Stay Safe and Regards
Adithi
