### 2020-09-07 #2630 @adithirgis
Started review [here.](https://github.com/VincentAlcazer/StatAid/issues/1)
