### 2020-09-07 VincentAlcazer/StatAid#1 @VincentAlcazer
Dear @adithirgis, thank you very much for your review and your comments. I will take the time to answer point by point and correct the issues you raised.
