### 2020-09-08 c2c7a86
v1.01
First revision according to @adithirgis comments
