+++ R/mod_Explo_table_des.R
@@ -4,138 +4,141 @@
 #'
 #' @param id,input,output,session Internal parameters for {shiny}.
 #'
-#' @noRd 
+#' @noRd
 #'
-#' @importFrom shiny NS tagList 
-mod_Explo_table_des_ui <- function(id){
+#' @importFrom shiny NS tagList
+mod_Explo_table_des_ui <- function(id) {
   ns <- NS(id)
   tagList(
-    
-    column(10,
-           h4("Analysis informations"),
-           htmlOutput(ns("analysis_info"))
-           
+    column(
+      10,
+      h4("Analysis informations"),
+      htmlOutput(ns("analysis_info"))
     ),
-    
-    column(10,
-           DT::DTOutput(ns("Table_des"))
-    ),
-    column(2,
-           absolutePanel(width = 200, right = 20, draggable = T,
-                         style = "opacity: 0.85",
-                         wellPanel(
-                           selectInput(ns("Group"), label =("Groups"),
-                                       multiple = F,selected = NULL,
-                                       ""),
-                           checkboxInput(ns("Include_na"), 
-                                         label="Include NA in groups",value=F),
-                           selectInput(ns("Variable"), label =("Variable(s)"),
-                                       multiple = T,selected = "All",
-                                       "All"),
-                           checkboxInput(ns("Percent_by_row"), 
-                                         label="Percent by row",value=F),
-                           radioButtons(ns("Adjust_pval"), "Adjust p-values:",
-                                        choices = list("No"="none", "Benjamini Hochberg (FDR)"="fdr", "Holm"="holm","Bonferroni"="bonferroni"), 
-                                        selected = "none"),
-                           actionButton(ns("Run_analysis"),"Run analysis")
-                           
-                         )
-           )#Absolutepanel
-    )#Column
-  )
-    
 
+    column(
+      10,
+      DT::DTOutput(ns("Table_des"))
+    ),
+    column(
+      2,
+      absolutePanel(
+        width = 200, right = 20, draggable = T,
+        style = "opacity: 0.85",
+        wellPanel(
+          selectInput(ns("Group"),
+            label = ("Groups"),
+            multiple = F, selected = NULL,
+            ""
+          ),
+          checkboxInput(ns("Include_na"),
+            label = "Include NA in groups", value = F
+          ),
+          selectInput(ns("Variable"),
+            label = ("Variable(s)"),
+            multiple = T, selected = "All",
+            "All"
+          ),
+          checkboxInput(ns("Percent_by_row"),
+            label = "Percent by row", value = F
+          ),
+          radioButtons(ns("Adjust_pval"), "Adjust p-values:",
+            choices = list("No" = "none", "Benjamini Hochberg (FDR)" = "fdr", "Holm" = "holm", "Bonferroni" = "bonferroni"),
+            selected = "none"
+          ),
+          actionButton(ns("Run_analysis"), "Run analysis")
+        )
+      ) # Absolutepanel
+    ) # Column
+  )
 }
-    
+
 #' Explo_table_des Server Function
 #'
-#' @noRd 
-mod_Explo_table_des_server <- function(input, output, session, r){
+#' @noRd
+mod_Explo_table_des_server <- function(input, output, session, r) {
   ns <- session$ns
- 
-  
+
+
   data <- reactive(r$test$data)
-  
+
 
   ## Parameters
-  
+
   observe({
     updateSelectInput(
       session,
       "Group",
-      choices=c(names(data()[-1][sapply(data()[,-1], class) %in% c("factor","character")])))
- 
[…]
