### 2020-09-08 VincentAlcazer/StatAid#1 @VincentAlcazer
Dear @adithirgis ,

Thank you for your review and your positive feedback. Your comments helped me to improve the software with a new version I just uploaded to GitHub (1.01). Unfortunately, I did not manage to reproduce some errors you get. Here is a point-by-point answer. I hope we will manage to find where these errors came from.

Thank you again for your time.

Best regards,

Vincent

1. The table area have been moved down as suggested, with the addition of a “frequent issue” section.

2. This is working for me both in local and on the website (see picture for expected behavior). I am trying to see where this error came from: Have you tried to perform the analysis on a fresh launch? Did you load another dataset before loading again the base dataset? Are your packages up to date?
![2](https://user-images.githubusercontent.com/63184395/92458201-5f6c3600-f1c5-11ea-8198-240b984925e4.PNG)

3. Same comment as for the 2nd point (see picture for expected result).
![3](https://user-images.githubusercontent.com/63184395/92458229-685d0780-f1c5-11ea-90a5-9d96d6672bca.PNG)

4. Additional description and guidelines for these points have been added at the end of the Readme.md.

5. Thank you for this suggestion. This is indeed an important point and I am currently considering writing a complete tutorial for the software. A short description of the dataset has been added.

6. To my knowledge, no free open-source software for data analysis exist. Of course, a lot of alternative solutions exist but they are else private software or shareware, or partial webtools with few options. I made a quick additional search on the web and did not manage to find direct concurrent in the same format as StatAid (if you have anything in mind please let me know so I can compare them).

7. Few papers where StatAid was useful have been published yet. The only one would be Clin Lung Cancer. 2020 May 8:S1525-7304(20)30136-4. doi: 10.1016/j.cllc.2020.04.013, where the software has been used for data exploration but not cited (I was added as an author). The PhD and medical thesis for which StatAid was more thoroughly used are else not going to be published, or with an article in preparation and I hope StatAid will be citable before the submission of these latter works.

8. The styler package has been used to reformat the whole code in a cleaner way.

9. I already run this tool before submission and ran it again with no warnings (only 3 notes). I wonder if a location issue could cause this and be potentially related to the 2nd and 3rd issue. Could you please paste me the warnings you had?

10. The recommended correction has been added.

11. From what I understood, the paper.md has to be short and should not contain API functionality (from JOSS recommendations: “Given this format, a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.”).
This point could probably be related to 5, and I am currently working on a more complete tutorial/manual for the software.

12. The rhub::check_for_cran() test has also been run, with 4 notes and no warnings. Do you have a particular test in mind you would like to see here?

13. The decimal separator for the Iris dataset is a period (instead of a comma for the original dataset). You have to make the change in the data loading menu so numerical variables are not considered as categorical ones. I added a section in “frequent issues” for this as this could be a regular issue.
