### 2020-09-10 VincentAlcazer/StatAid#2 @nistara
JOSS Review
Hi @VincentAlcazer,

Thank you for your submission to JOSS and for inviting me to review. I think StatAid is a great app, and I wish I had come across it sooner!!

Here are my comments and suggestions:

#### Regarding the app
1. Introduction
    a. The text bleeds off of the grey content background (noticed both on Chrome and Safari browsers):

2. Data loading
In addition to the sample dataset already provided, I was able to explore the app using this [penguin dataset](https://github.com/allisonhorst/palmerpenguins/blob/master/inst/extdata/penguins_raw.csv) as well. It uploaded and worked well.

    a. Consider emphasizing the sentence describing the file types supported by your app, so it stands out for new users: e.g. make the *"...your dataframe should be saved as a tab (.txt/.tsv), comma or semicolon (.csv) delimited file"* bold or italic or add some sort of emphasis on this key point!

    b. Regarding the example dataset: In the Readme or vignette/tutorial, consider giving a slightly more detailed explanation of each variable. Also, consider your target audience....if it's worldwide beyond France, you should think about changing the variable name "Sexe" to "Sex".

    c. Enable horizontal scrolling in the table view because columns are otherwise cut off to the right:

    They remain cut off even when I zoom out a lot:

3. Data exploration
   a. In "Categorical variables" exploration, there is a very useful "Summary table". It's showing the breakdown by only one variable, however, and not both variables plotted (for e.g. no FAB in the screenshot below):

    b. In the Descriptive table, I think it's great and I appreciate you put in a short explanation of the methods used. Suggestions for the methods paragraph, below, are:

    *Methods : Categorical variables are expressed as n (%) and compared with the Chi-squared test or its non-parametric alternative Fisher's test with simulated p-values. Numerical variables are expressed as mean **(standard-derivation)** or median **[IQR]** and compared with **else** Welch's t-test (or its non-parametric alternative Wilcoxon's rank-sum test) or ANOVA (or its non-parametric alternative Kruskal-Wallis test) where appropriate. Variables with >80% missing are removed from the analysis. P-values are not adjusted.*

    Change **standard-derivation** to **standard deviation**,  **IQR** to **interquartile range**, and **else** to **either** above.

    c. In custom graphs, the text might be overlaid on the graph itself...so possibly have an option to increase Y-axis limits? I tried increasing the Y axis limits, but didn't see a change in the plot:

4. Paired-data analysis section
    a. In the data format, you state: *"each row correspond to an observation (sample/patient), each selected variable/column to a different timepoint of the same variable (e.g. Concentration_day0, concentration_day15, concentration_day30...)"*

    The example dataset you provided doesn't have columns corresponding to the names above. I know users will have their own datasets with paired/time-dependent columns, but for the sake of a good explanation, your sample dataset should have columns which can be used in this case, and whose column names are identifiable in the example you give above. Doing so will be much more intuitive for new users.

5. Univariate analysis
    a. I could run the linear model, but not GAM or LOESS in the Continuous Univariate analysis section:

    The error I get is:
    ```
    Warning: Error in eval: object 'Age' not found
    ```

    I'll check to see if there's an outdated package at my end causing this issue. If so, others might comes across this issue as well.

6. Points on the paper, documentation, and tests
    a. Your statement of need says *"The package provides all the tools necessary to perform a complete data analysis in an intuitive ready-to-use graphical interface. Any user with no coding skill or no software with paid-license available can easily perform all the steps of a good statistical analysis, from data-exploration/quality check to multivariate modeling."*

    I think this is an excellent application that provides essential and core statistical analytical tools, however, I don't think it provides *"all the tools for a complete data analysis"*. I'm sure you'll agree that there are many different types of analyses that suit different situations (and have various assumptions), for e.g. mixed-effects models might in many cases be more appropriate than standard linear regression models in health data analyses. It's great that you later specify that users can request new features (further adding emphasis to the point above that not all tools are included) - I think this evolving aspect, in combination with the existing features make it a solid package. I therefore recommend you to update your Statement of need to reflect your package's use better.

    While I see how useful this application can be (and believe it's been used as you say), I also agree with Adithi regarding prior use of this software: if you state that it's been used by researchers and doctors, it helps to have a citation, else it's not verifiable.

    Btw, regarding other apps that do the similar things, I came across other shiny apps that provide various aspects of statistical analysis and visualization, for e.g. https://pkgs.rsquaredacademy.com/, but your app is unique in that it incorporates various steps of the analytical workflow in one package itself.

    b. In both the paper and README, under the features section, there is the point *"Paired data analysis (case-control studies, repeated measures)"*. Case control studies are a type of observational study and it's not necessary that paired data analysis will always be used for this kind of study (for e.g., even logistic regression can be used, depending on the type of case-control study, matched pairs vs not, etc).

    c. One important thing about the analyses you've included: there are no references about functions from various packages (or base R) that are used for statistical analyses, nor any references about the statistical methods themselves. The Contact/About page has the full package list though, and it would be a good place to add summarized information about specific functions used to perform various tests: e.g. glm(), coxph(), gam(), cor.test(), and any others you used. Your users can then look these functions up for further references and also get an idea of default function options and any other pertinent information. This could be crucial for people considering using your package for their research and publications.

    d. In line with Adithi's suggestion, I recommend adding more examples to your documentation. Beyond making it easier for users to grasp what your package is doing and enabling them to see what's possible, it is also a requirement for JOSS submission that we have to check for. The addition of a short description of the dataset helped immensely, and adding more examples to the README would be useful while you work on your detailed tutorial on the side.

    e. Regarding tests, you've got three tests pertaining to the app itself (as seen in [test-golem-recommended.R](https://github.com/VincentAlcazer/StatAid/blob/master/tests/testthat/test-golem-recommended.R) and when `devtools::test()` is run). There aren't any tests for the functions in [Functions.R](https://github.com/VincentAlcazer/StatAid/blob/master/R/Functions.R) themselves. It's recommended to have tests for functions in your package, especially as it starts adding more features and becomes more complex.  They'll also lend further credibility to your package. A useful resource for this would be https://mastering-shiny.org/scaling-testing.html. I suggest adding a few more tests, as also  recommended by Adithi, while understanding that it's an iterative process which will evolve even after the review is finished.

7. I've made minor typographical corrections to your app interface, documentation and paper, which you can see via my pull request. It was easier to do it like that than mention them separately.

8. Some suggestions which you could keep in mind as your package evolves. You don't need to incorporate them now, but it's food for thought:

    a.  X axis labels can go outside of plot margins if the labels are really large (for e.g. with the penguin dataset). Ideally the labels won't be so big, but just bringing this to your notice.

    b. Using color palettes suitable for the colorblind people amongst us. Right now the plots will appear as below to people with Deuteranopia (red-green colorblindness, which is the more common form):

    ![Deuteranopia3](https://user-images.githubusercontent.com/12116051/92675059-713bff00-f2d3-11ea-926c-00e143c6947d.png)

    ![Deuteranopia2](https://user-images.githubusercontent.com/12116051/92675035-66816a00-f2d3-11ea-9dd4-3e8bbaafcd29.png)

    It's difficult to choose a colorblind friendly palette when there are many categories, but as seen in the topmost image, a palette could be chosen such that it is more easily distinguishable when there are fewer categories.

    c. A way to export the tables and graphs would be great.

    d. The figure options menu on the right overlaps with the plots depending on the zoom level:
    At 100% zoom:

    At 90% zoom:

Please let me know if you have any questions on my comments above. I think this app is really good, and I can definitely see people using it to explore and work with their data.

Merci,
Nistara
