+++ paper.md
@@ -22,10 +22,10 @@ affiliations:
 
 # Summary
 
-Data analysis is a crucial step in every research project in life science. Every clinician or researcher is one day faced to the need of performing statistical analysis. However, few free accessible solutions exists to date and most of the reference software need a paid license. R is a free langage allowing to perform statistical analysis [@RCoreTeam:2017].
-While R environment is very powerful, its learning curve can be very steep at the beginning, especially for people with no previous coding skill or few time to learn. A graphical user interface has already been provided as an independant package, but its features are limited for medical and applied life science studies and its usage remains difficult and not intuitive for new users [@Fox:2005]. 
+Data analysis is a crucial step in every research project in life science. Every clinician or researcher is one day faced to the need of performing statistical analysis. However, few free accessible solutions exist to date and most of the reference softwares need a paid license. R is a free language allowing one to perform statistical analysis [@RCoreTeam:2017].
+While the R environment is very powerful, its learning curve can be very steep at the beginning, especially for people with no previous coding skills or less time to learn. A graphical user interface has already been provided as an independent package, but its features are limited for medical and applied life science studies and its usage remains difficult and unintuitive for new users [@Fox:2005]. 
 
-`StatAid` is a free open-source software provided as an R package allowing clinicians and researchers to perform statistical analysis through an intuitive graphical interface. It has been developed  using the Shiny package [@Chang:2020]. Golem has been used for package compilation and deployment [@Guyader:2020].
+`StatAid` is a free open-source software provided as an R package allowing clinicians and researchers to perform statistical analysis through an intuitive graphical interface. It has been developed using the Shiny package [@Chang:2020]. Golem has been used for package compilation and deployment [@Guyader:2020].
 
 The software guide the users through all the steps of a data analysis, including multiple features:
 
@@ -35,24 +35,24 @@ The software guide the users through all the steps of a data analysis, including
 
 - Publication-ready graph customization 
 
-- Paired data analysis (case-control studies, repetead measures)
+- Paired data analysis (case-control studies, repeated measures)
 
 - Univariate analysis and models for continuous and categorical outcome: Correlation, linear and logistic regression 
 
-- Univariate analysis and models for time-dependant outcome: Kaplan-Meier curves and cox regression 
+- Univariate analysis and models for time-dependent outcome: Kaplan-Meier curves and cox regression 
 
-- Multivariate analysis and models for continuous, categorical and time-dependant outcomes
+- Multivariate analysis and models for continuous, categorical and time-dependent outcomes
 
-Its user-friendly interface can guide any clinician or researcher even with no previous software experience for its statistical analysis. In addition to a local version, a ready-to-use online version [(https://vincentalcazer.shinyapps.io/StatAid/)](https://vincentalcazer.shinyapps.io/StatAid/) is also alvailable, providing access to `StatAid` everywhere even on hospital / research center computers where no external software can be installed.
+Its user-friendly interface can guide any clinician or researcher even with no previous software experience for its statistical analysis. In addition to a local version, a ready-to-use online version [(https://vincentalcazer.shinyapps.io/StatAid/)](https://vincentalcazer.shinyapps.io/StatAid/) is also available, providing access to `StatAid` everywhere, even on hospital/research center computers where no external software can be installed.
  
 
 # Statement of need 
 
-`StatAid` is an R package designed to fit the needs for every-day statistical analysis in science. The package provides all the tools necessary to perform a complete data analysis in an intuitive ready-to-use graphical interface. Any user with no coding skill or no software with paid-license available can easily perform all the steps of a good statistical analysis, from dataexploration/quality check to multivariate modeling. The online version of `StatAid` render it accessible everywhere, even on computers with restricting software installation policies.
[…]
