+++ README.Rmd
@@ -19,16 +19,16 @@ knitr::opts_chunk$set(
 [![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
 <!-- badges: end -->
 
-`StatAid` is a free open-source software provided as an R package allowing clinicians and researchers to perform statistic analysis through an intuitive graphical interface. It has been developed with the R software, using the [Shiny package](https://shiny.rstudio.com/). [Golem](https://github.com/ThinkR-open/golem) has been used for package compilation and deployment.  
+`StatAid` is a free open-source software provided as an R package allowing clinicians and researchers to perform statistical analysis through an intuitive graphical interface. It has been developed with the R software, using the [Shiny package](https://shiny.rstudio.com/). [Golem](https://github.com/ThinkR-open/golem) has been used for package compilation and deployment.  
 
- The software guide the users through the steps of a good data analysis, including multiple features:
-<ul><li> Exploratory data analysis: distribution, count, missing-values and outliers check  </li>
+ The software guides the users through the steps of a good data analysis, including multiple features such as:
+<ul><li> Exploratory data analysis: distribution, count, missing-values and outliers check</li>
 <li> Descriptive analysis, simple comparative analysis and publication ready 'table 1' output </li>
-<li> Publication-ready graph customization  </li>
-<li> Paired data analysis (case-control studies, repetead measures) </li>
-<li> Univariate analysis and models for continuous and categorical outcome: Correlation, linear and logistic regression  </li>
-<li> Univariate analysis and models for time-dependant outcome: Kaplan-Meier curves and cox regression </li>
-<li> Multivariate analysis and models for continuous, categorical and time-dependant outcomes </li></ul>
+<li> Publication-ready graph customization</li>
+<li> Paired data analysis (case-control studies, repeated measures) </li>
+<li> Univariate analysis and models for continuous and categorical outcome: Correlation, linear and logistic regression</li>
+<li> Univariate analysis and models for time-dependent outcome: Kaplan-Meier curves and cox regression </li>
+<li> Multivariate analysis and models for continuous, categorical and time-dependent outcomes </li></ul>
 
 # Getting started
 
@@ -43,9 +43,9 @@ You can install the development version from [GitHub](https://github.com/Vincent
 ```{r Github install }
 
 # install.packages("remotes")
-#remotes::install_github("VincentAlcazer/StatAid")
+# remotes::install_github("VincentAlcazer/StatAid")
 
-#StatAid::run_app()
+# StatAid::run_app()
 ```
 
 
