### 2020-09-10 VincentAlcazer/StatAid#2 @nistara
Additional comment: the references' capitalizations for R need to be fixed.
