### 2020-09-10 VincentAlcazer/StatAid#2 @VincentAlcazer
Dear @nistara ,

Thank you very much for your review and feedback. I am currently away from home and I will take the time to respond point by point as soon as possible.

Best regards,

Vincent
