### 2020-09-10 VincentAlcazer/StatAid#1 @adithirgis
Hello @VincentAlcazer !

Thanks. Sorry for the late response.

- I see that 1, 4, 10,  and 13 have been implemented. Thanks!
- For 2 and 3, I tried with three Windows 10 system and RStudio 3.5, 3.6 and 4.0 with all updated packages (for packages with specific version please add that in your [DESCRIPTION](https://github.com/VincentAlcazer/StatAid/blob/master/DESCRIPTION)), new session with just the example data to be tested. I still find the same error, yes the online version works perfectly well. (though comment 9 has been resolved)
```
Warning in chisq.test(table) :
  Chi-squared approximation may be incorrect
Warning: Error in : Tibble columns must have compatible sizes.
* Size 8: Existing data.
* Size 4: Column `ELN2017`.
i Only values of size one are recycled.
  146: <Anonymous>
`geom_smooth()` using formula 'y ~ x'
Warning in predict.gam(x, newdata, type = type.predict) :
  not all required variables have been supplied in  newdata!
Warning: Error in eval: object 'Age' not found
  176: <Anonymous>
```
I also tried the [RStudio cloud](https://rstudio.cloud/) to test this, I get the same errors 2 and 3, pasting 2 and error from RStudio cloud -
![image](https://user-images.githubusercontent.com/63432077/92633932-25984e00-f2f1-11ea-94a9-fcd300e46d22.png)
![image](https://user-images.githubusercontent.com/63432077/92634164-7f007d00-f2f1-11ea-9edd-7afb828dee80.png)
I would like to know what is wrong here so that other users do not face the same issue.
- For 5, I would wait for the README.md to complete to wrap up the review process.
- Please add 6 in the paper.md that author is not aware of any similar free, open-source software like StatAid (of course in the right way).
- For 7, so StatAid was used but not cited. I wanted to know to verify it. Thanks.
- For 9, I used the new version you uploaded (updated the packages, and started a new session), now I find just three notes. Great! Sorry, I did not do that earlier.
- For 12, I would suggest [testthat](https://testthat.r-lib.org/) / [shinytest](https://rstudio.github.io/shinytest/articles/shinytest.html) apart from what Nistara mentioned for testing shiny apps, I do see a tests folder made using testthat with the default test file. Please add test unit for automatically testing your application. Tests for functions will also be required. Examples from JOSS - https://github.com/michaellevy/gwdegree/tree/master/tests/testthat, https://github.com/PsyChiLin/EFAshiny/tree/master/tests/testthat, https://github.com/OttaviaE/DscoreApp/tree/master/tests, https://github.com/pmcharrison/psychTestR/tree/master/tests/testthat.

Please treat the following as minor comments / optional-

- For 8, I would still recommend using the resources shared before since some of the code cleaning does not happen using that particular project as seen [here](https://github.com/VincentAlcazer/StatAid/blob/master/R/Functions.R). (the characters in a line seem way more than 120 characters).
![image](https://user-images.githubusercontent.com/63432077/92630976-4e6b1400-f2ee-11ea-9082-8a07e856c02a.png)
- For 11, I was wondering if a brief descriptive explanation of analysis that StatAid can achieve (Since I see a lot of analysis can be done using this.) It would help get more visibility.
- I think, I too agree with Nistara's view about the colour palette. It will be good step.
- If there are options to export table it would be great. I could save the graphs but I could not export the tables (if this helps: https://stackoverflow.com/questions/45270312/downloading-multiple-reactive-objects-containing-data-tables).

Feel free to let me know if you have questions.
All the best! I will surely try to use this for air quality data! Thanks a lot.
Regards and Stay safe
Adithi
