### 2020-09-23 #2630 @VincentAlcazer
Dear @mikldk,
I  am currently working on the issues raised by the reviewers and their comments are really helping me improving the software.
Due to the pandemic evolution in France, I am also quite busy at the hospital with only few moments to work on StatAid. However, I already corrected the main issues reported and I hope I'll be able to provide my complete answer soon.
I am very sorry for this delay and will do my best to proceed through the reviewing in time.
Best regards,
