### 2020-10-05 db2353f
v1.02
Reviewing process - Second round of corrections
