+++ R/mod_Model_surv.R
@@ -53,7 +53,10 @@ mod_Model_surv_ui <- function(id) {
                 choices = list("Cox" = "cox"),
                 selected = "cox"
               ),
-              actionButton(ns("Run_analysis"), "Run Analysis")
+              actionButton(ns("Run_analysis"), "Run Analysis"),
+              p(),
+              downloadButton(ns("download"), "Download table (.tsv)")
+              
             )
           ) # Absolutepanel
         )
@@ -155,6 +158,16 @@ mod_Model_surv_server <- function(input, output, session, r) {
       default_theme +
       labs(title = "Univariate analysis - Cox model", y = "Hazard ratio [95%CI]", x = "") +
       theme(legend.position = "right")
+  )
+  
+  # Download table
+  output$download <- downloadHandler(
+    filename = function() {
+      paste("Univariate_cox_table.tsv")
+    },
+    content = function(file) {
+      write.table(regression_table_df()[, -7], file, row.names = FALSE, sep = "\t", quote = F)
+    }
   )
 }
 
