+++ dev/03_deploy.R
@@ -17,6 +17,7 @@
 ## Check the package before sending to prod
 devtools::check()
 rhub::check_for_cran()
+devtools::test()
 
 # Deploy
 

[…]
