### 2020-10-05 VincentAlcazer/StatAid#2 @VincentAlcazer
Dear @nistara ,

Thank you again for your complete review of my work and for your relevant comments. I just uploaded a corrected version (1.02) of StatAid on github and Shinyapps.io. I managed to take into account most of your comments and it really helped me improving the app. However, there is still some points (mostly in line with @adithirgis comments) that I could not manage to modify yet, but I will work on it for the future corrections:
-	The GAM / LOESS issues, also reported by @adithirgis: I could not manage to reproduce the error and tried to add the mgcv package as a systematic loading. Please tell me if it resolved the error for you.
-	The detailed tutorial/guide is still work in progress.
-	The need for more test, as commented in my answer, will also take some time.

You will find attached a detailed list of the corrections. Thank you again for your time.

Best regards,

Vincent

1. This has been fixed.

2.a. The sentence has been put in bold.
2.b. The variable name has been changed. A short description of the dataset has been added directly on the app. A more complete tutorial/description is currently under preparation for the readme, as suggested by @ adithirgis. A more complete description of the dataframe will be included as suggested.
2.c. The table has been fixed with a horizontal scrollbar.

3.a. Variable count has been added to the table.
3.b. The changes have been made.
3.c. The Y-limits control slider has been fixed.

4.a. Three simulated “concentration” data columns have been added to allow the testing of this feature with the base dataset.

5.a. @adithirgis reported the same issue. I could not manage to reproduce the error even after having installed StatAid on a new computer with no R nor Rstudio installation. I will try to see on a computer with a previously installed R version. I tried to add the mgcv package as a systematic import to see if this could correct the issue.

6. a. I agree with you and have modified the paper consequently. I am no longer talking about a “complete” data analysis, but kept “all the steps of a data analysis”, talking here more about the different steps (from data encoding, quality and distribution check to analysis itself) than features exhaustivity.
As there is currently no evidence of StatAid use (and as this point as also been noted by @adithirgis), the sentence “It has already been used by multiple researchers and students for their PhD or medical doctors for their thesis” has been removed.
b. This point has been modified as suggested: “- Paired data analysis (Repeated measures...)”
c. The package list in the contact section has been complete modified to include a more complete description of packages and base R function usage.
d. In line with 2.b, this will be included in the complete tutorial I am working on.
e. Thank you for this reference from mastering-shiny.org. I read with interest the test section, together with multiple tutorials. I built StatAid in a way that each loaded dataset is formatted in a tidy way working with the functions I designed. The limiting step is currently the loading and formatting (if the data load correctly, the function will work), that’s why I find it hard to see which specific test for my functions I could use (according to test_that: I hardly see the point to add an “expect_equal” or “expect_error” in my functions). I took note that it is an important for enhancing contributions and code sharing and will try to figure out what relevant test I could add here.

7. All the correction have been merged. Thank you very much for this diligent proofreading.

8. a. Thank you for this suggestion. For long variable name, I usually invite to set only the external legend on the right or bottom for example and hide it on the x-axis. I will evaluate the necessity to add a function to cut variable names longer than n characters if users really want to see it on the x axis. I could also extend the angle to 90° so it does not longer go outside of the margins.
b. This is indeed an important point. I have already added different color palettes for the custom graph section, but all the exploration modules are plotted with the base ggplot2 colors. I will work on a “colorblind” mode for these sections in the right panel.
c. A download button has been added for all the tables. Graphs can already be downloaded with a right click. I would like to add a download button that allows to modify the size and dpi of the graph, but I have not found any module that provide such feature for the moment.
