### 2020-10-05 VincentAlcazer/StatAid#1 @VincentAlcazer
Dear @adithirgis,

Thank you for your answer and your patience. I just uploaded a new version (1.02) integrating @nistara comments and some of your remaining reported issues. Three big points are still pending (the GAM/LOESS issue that I tried to correct, the guide/tutorial and the need for more test that are still work in progress) and I hope I will manage to provide a full correction soon.

Thank you again for your time and your help improving StatAid.

Best regards,

Vincent

2-3. For the GAM/LOESS issue, I tried to force the loading of the mgcv package at the app start: could you please tell me if you still have the issue?
5. The readme file with the tutorial/guide is still work in progress. I will let you know when it will be ready.
6. A sentence has been added in the paper.md: “To my knowledge, no free open-source software directly designed for researchers with an intuitive interface and a collaborative/evolving environment has been proposed yet.”
7. I removed the statement of previous use as I have no citations/proof to provide.
12. The testing part has also been commented by @nistara’s review. As I explained, I hardly see what kind of test I could add to my functions for the moment in the way I designed them (i.e. each function take the dataframe and the columns as input to produce the results, making the dataframe loading the limiting step). I am currently working on testing solutions and will try to propose relevant test with testthat for the next corrections.

Minor comments:
-	I will make a full review of the style according to your resources once major corrections will be done, using styler as a base and controlling each file manually.
-	11 will be integrated in the tutorial/guide.
-	A download button has been added for data tables.
