+++ paper.bib
@@ -10,7 +10,7 @@
 
 @article{Fox:2005,
 	author = {Fox, J.},
-	title = {The R Commander: A Basic-Statistics Graphical User Interface to R},
+	title = {The {R} Commander: A Basic-Statistics Graphical User Interface to {R}},
 	journal = {Journal of Statistical Software},
 	volume = {14},
 	number = {9},
@@ -20,7 +20,7 @@
 }
 
 @article{Chang:2020,
-    title = {shiny: Web Application Framework for R},
+    title = {shiny: Web Application Framework for {R}},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
     year = {2020},
     note = {R package version 1.4.0.2},
