### 2020-10-05 VincentAlcazer/StatAid#2 @VincentAlcazer
PS: The references' capitalization have been fixed!
