### 2020-10-05 VincentAlcazer/StatAid#1 @adithirgis
Hi @VincentAlcazer!

Thanks for the response! I will take a look again and wrap the review process once the tutorial is updated.

Take care,
Adithi
