### 2020-10-05 8557168
1.02
