+++ R/mod_Explo_paired.R
@@ -51,7 +51,7 @@ mod_Explo_paired_ui <- function(id) {
             radioButtons(ns("Error"), "Error bars (Barchart)",
               choices = c(
                 "Hide" = "hide",
-                "Standard derivation" = "sd",
+                "Standard deviation" = "sd",
                 "Standard error" = "se",
                 "95% CI" = "IC95"
               ),
