### 2020-10-06 VincentAlcazer/StatAid#1 @adithirgis
Thanks again for all the changes, the app has been improved.

I see that all the comments have been implemented and just 3 of them still remain.

1. The issues regarding running different models - I would try to run the app on [RStudio Cloud](https://rstudio.cloud/plans/free) and check what happens there. I tried to use the new version of the application on my system and on cloud, I still get the same error.
2.  I will wait for the README.md file to be completed for wrapping the review process.
3. To add tests to the package -  Its mostly based on how the app is supposed to work. I can say with experience of using `shinytest` that the tests are to check the working after you make changes to your app in any way. For eg:- How would your graphs change when you use different variables and if they are correct then add the test. Make changes to your app (eg- adding buttons or changing variables). If there is any significant changes in the app then the tests will change and you can update them, if expected. If you do not expect the changes then fix the bugs. (Sorry if I am not able to portray my thoughts correctly.) I think this blog might help - https://blog.rstudio.com/2018/10/18/shinytest-automated-testing-for-shiny-apps/.

Regards,
Adithi
