### 2020-10-07 VincentAlcazer/StatAid#2 @nistara
Thanks @VincentAlcazer ! I'll go through the edits and get back to you shortly.
