### 2020-10-08 eaaec55
v1.03
Bug correction (GAM/LOESS and descriptive table)
