+++ dev/02_dev.R
@@ -68,7 +68,7 @@ usethis::use_data_raw(name = "my_dataset", open = FALSE)
 
 ## Tests ----
 ## Add one line by test you want to create
-usethis::use_test("app")
+usethis::use_test("functions")
 
 # Documentation
 
