+++ R/mod_Model_cat_multi.R
@@ -148,7 +148,7 @@ mod_Model_cat_multi_server <- function(input, output, session, r) {
     regression_table_df() %>%
       ggplot(aes(x = multiv_graph, y = `Odds Ratio`, color = `X Variables`)) +
       geom_pointrange(aes(ymin = CI95_low, ymax = CI95_high), size = 1) +
-      geom_hline(yintercept = 0, linetype = "dashed", size = 1) +
+      geom_hline(yintercept = 1, linetype = "dashed", size = 1) +
       coord_flip() +
       default_theme +
       labs(title = "Multivariate analysis - Linear model", y = "Beta coefficient", x = "") +
