### 2020-10-08 VincentAlcazer/StatAid#2 @VincentAlcazer
Dear @nistara ,

I hope your are doing well.

I wanted to inform you that I managed to correct the three remaining points of the review. You will find in the updated v1.03:

- A correction of the GAM/LOESS and descriptive table issue
- A quick-start user guide with usage example (can be found in the readme or here: https://github.com/VincentAlcazer/StatAid/blob/master/STATAID_QUICK_START_USER_GUIDE.pdf)
- Implemented test for each function of the package.

Thanks again for your time and comments. I'm looking forward to your feedback.

Best regards,

Vincent
