### 2020-10-08 VincentAlcazer/StatAid#1 @adithirgis
Hi @VincentAlcazer!

Sorry for getting back late.

Thanks again for all the changes, the app has been improved a lot.
I see that all the comments have been implemented. I will wrap the review process soon.

Could you please share what changes did you make to fix the GAM/LOESS analysis (it works on both Cloud and local system now!)?
Thanks for such a wonderful user guide!
Tests also seem to be added for functions! Please continue adding the tests to the test folder to improve the application time and again.

Regards,
Take care!
Adithi
