### 2020-10-09 VincentAlcazer/StatAid#1 @adithirgis
Hi @VincentAlcazer!

Congratulations for the amazing application. I will wrap the review process now.

All the best and take care!
Adithi
