### 2020-10-09 VincentAlcazer/StatAid#1 @VincentAlcazer
>
>
> Hi @VincentAlcazer!
>
> Sorry for getting back late.
>
> Thanks again for all the changes, the app has been improved a lot.
> I see that all the comments have been implemented. I will wrap the review process soon.
>
> Could you please share what changes did you make to fix the GAM/LOESS analysis (it works on both Cloud and local system now!)?
> Thanks for such a wonderful user guide!
> Tests also seem to be added for functions! Please continue adding the tests to the test folder to improve the application time and again.
>
> Regards,
> Take care!
> Adithi

The broom::augment function was not supporting the fit object from GAM so I had to manually recode the augment dataframe output by my function. However, I still do not understand why it worked on shinyapps.io and on my local computer...! Thanks for providing me the Rcloud resource that helped me fix this!
