### 2020-10-09 VincentAlcazer/StatAid#2 @nistara
Dear @VincentAlcazer ,

Thanks for incorporating the edits even as you're busy with the ongoing pandemic. Hope you're doing well and are safe.

**Comments on the edits:**
I'm confirming that the GAM/LOESS models are working for me, and that the quick-start user guide is accessible and helpful for people to get acquainted with your app.

_6 b. This point has been modified as suggested: “- Paired data analysis (Repeated measures...)”_
I somehow still see this as _Paired data analysis (case-control studies, repeated measures)_ in the latest paper.md and the app Intro page. In addition, there's a minor typo in this line in the app Intro page (repeated's spelling is wrong).

It'd be fine if you specified "matched case-control studies" above instead of just case-control.
(As an aside, you might be interested in this paper: https://www.bmj.com/content/352/bmj.i969)

_6 e. Thank you for this reference from mastering-shiny.org. I read with interest the test section, together with multiple tutorials. I built StatAid in a way that each loaded dataset is formatted in a tidy way working with the functions I designed. The limiting step is currently the loading and formatting (if the data load correctly, the function will work), that’s why I find it hard to see which specific test for my functions I could use (according to test_that: I hardly see the point to add an “expect_equal” or “expect_error” in my functions). I took note that it is an important for enhancing contributions and code sharing and will try to figure out what relevant test I could add here._

It's great you added the tests! It's a good start and something that should help you going forward as you incorporate additional functionalities and tests to your app. I did get one error when I ran `devtools::test()` on your tests. It was for the `regression_table_multi_lm` test:

```
res$`Beta Coeff.`[1] not equal to 1.3.
1/1 mismatches
[1] 1.34 - 1.3 == 0.04
```
You might want to change
```
expect_equal(res$`Beta Coeff.`[1],1.3)
```
to
```
expect_equal(res$`Beta Coeff.`[1],1.34)
```
Or would that break the test at your end?

Also, in Paired-data analysis Methods, I assume the IC95 in **mean (IC95)** stands for "Intervalle de confiance". Considering the target audience of many countries, it would be preferable to change it to 95CI (confidence interval).

**The following are minor comments that won't affect the conclusion of my review, but are put here for your consideration:**

_3.c. The Y-limits control slider has been fixed._
Confirming that the slider's been fixed, however, the text still appears on the plots even after increasing the y-axis limits, and can be obscured by it.

6 b.
_As there is currently no evidence of StatAid use (and as this point as also been noted by @adithirgis), the sentence “It has already been used by multiple researchers and students for their PhD or medical doctors for their thesis” has been removed._

I still see this line even though I'm working on the latest file versions (to the best of my knowledge). I'm quite fine with you leaving it in since you've already clarified to us that it's not formally been cited, but it has been used.

_8 b. This is indeed an important point. I have already added different color palettes for the custom graph section, but all the exploration modules are plotted with the base ggplot2 colors. I will work on a “colorblind” mode for these sections in the right panel._

Sounds good! This might be helpful: https://github.com/clauswilke/colorblindr
I really appreciate that you have different color palettes to choose from. It's a great feature.

_8 c. A download button has been added for all the tables. Graphs can already be downloaded with a right click. I would like to add a download button that allows to modify the size and dpi of the graph, but I have not found any module that provide such feature for the moment._

There's no download option for the summary table in Data Exploration -> Categorical variables.

Finally, in your custom plot, the order of the variables in the plot is different from that of the legend. So visually, they're in reverse order.

Well, that's it from me!
Excellent work on this app, and I look forward to using it.

Best wishes,
Nistara
