### 2020-10-09 VincentAlcazer/StatAid#2 @VincentAlcazer
Dear @nistara ,

The paper.md was not up-to-date, sorry for this oversight. I think there was a conflict when I merged the pull request. It should now be online (correcting the matched case-control study and the sentences I had modified), together with the introduction correction and the 95CI (I tried to track all the French abbreviations but this one stood under my radars!).
The test has also been corrected.

Concerning the minor corrections: The X-variables order has been corrected. I will take the time to modify the plot function so that statistics will move with the y-axis limits.

Thank you again for your time and thorough your review.

Best regards,

Vincent
