### 2020-10-09 VincentAlcazer/StatAid#2 @nistara
Dear @VincentAlcazer,

Perfect! Thank you for incorporating the suggestions. My last minor comment is for you to update the Readme's Paired data analysis point so that it's consistent with your paper and app Intro.

I'm going to go ahead and close this issue. Thank you for inviting me as a reviewer: it was a pleasure to check this package out and explore its many functionalities.

Best wishes,
Nistara
