### 2020-10-09 #2630 @nistara
Dear @mikldk and @VincentAlcazer ,

I'm done with my review. Thank you for inviting me as a reviewer. This package is pretty awesome and should help many people explore various descriptive and statistical tools and visualizations with one tool!! I also appreciate the careful consideration give to what potential users might need to customize their analysis and plots.

Best wishes,
Nistara
