### 2020-10-12 #2630 @VincentAlcazer
Dear @mikldk ,

Thank you for giving me the opportunity to have my software reviewed by JOSS. I would like to thanks @nistara and @adithirgis who did an amazing job thoroughly reviewing StatAid and really helped me to improve the software.

Please let me know if there is anything else I should do now for the following submission process.

Best regards,

Vincent Alcazer
