### 2020-10-12 #2630 @mikldk
Thanks for @nistara and @adithirgis for their reviews.

@VincentAlcazer :

* How does this software compare to [iNZight](https://inzight.nz/)?
* Please make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
