### 2020-10-12 1fd626e
v1.1 release for JOSS publication
