### 2020-10-12 #2630 @VincentAlcazer
Dear @mikldk ,

I did not know iNZight which seems to be a pretty good software developed by a serious team. This software is however different from StatAid in multiple points:

-	StatAid is more user-friendly, giving the possibility to quickly plot publication-ready graph or tables. On the contrary, it offers less options than iNZight which provides a more complete environment to modify data and recode variables for example. The absence of data/variable modification in StatAid was a choice, allowing the implementation of variable check and controls before all tests and analyses. Hence, StatAid is pushing the author to understand its variables (the differences between numeric, categorical, time-dependent) and how its should be encoded so that StatAid recognize the good type of variable.
-	StatAid is really guiding the user for its data analysis. In that way, StatAid will make some choices and provide the method used (for example, StatAid will automatically choose the most appropriate statistics/test to use for some type of comparisons). Related to automatic variable detection, the software will help the user preventing bad test use/bad graph choice.
-	StatAid is providing graph customization solutions (such as famous journal colors palette (JCO, nature, NEJM…) or full graph legend modification option) which not seems to be implemented in iNZight.
-	Compared to iNZight, StatAid offer the possibility to perform survival analysis (Kaplan-meir curves, uni and multivariate cox analysis). Conversely, compared to StatAid, iNZight offer the possibility to perform modelisation with multiple responses.
-	StatAid is completely open-source and offer the possibility for users either to ask for a feature implementation or to provide an amelioration. This open-source aspect is also a security for people working with sensitive data (e.g. data from clinical trials / patients).
-	StatAid can be used online on shyniapps.io, which is an important feature for conditions where the user can’t install external software on the computer (which is the case in many hospitals for example).

All-together, I would say that StatAid is more convenient for quickly accessing publication-ready graph and table, really guiding the authors through their data analysis. iNZight seems to have a more complete solution (with data modification and recoding possibilities), to the cost of less user-friendly / guiding interface. Only StatAid is providing a time-dependent variable analysis option and has this evolving / contributive aspect.

I checked the paper.md and everything seems OK. I uploaded the 1.1 version on zenodo with the following DOI:

10.5281/zenodo.4081353

Thank you.

Best regard,

Vincent Alcazer
