### 2020-10-16 #2630 @adithirgis
Hi @mikldk and @VincentAlcazer,

I think it would good to add the strong points (for the benefit of the users) in the paper.
