### 2020-10-16 #2630 @nistara
I agree as well...these points provide a contrast for StatAid and also highlight some of its salient features. In addition to options for different statistical tests, I particularly liked the visualization options and the choice of different color palettes. A lot of consideration should go into analyses in general, and there could be models more suitable than those provided at the moment (e.g. mixed models), but I can see lots of people, including students, benefiting from exploring their data with this software and using it for their work. So yes, including the above considerations would definitely benefit the paper and the readers in understanding what the software offers.
