+++ README.Rmd
@@ -25,7 +25,7 @@ knitr::opts_chunk$set(
 <ul><li> Exploratory data analysis: distribution, count, missing-values and outliers check</li>
 <li> Descriptive analysis, simple comparative analysis and publication ready 'table 1' output </li>
 <li> Publication-ready graph customization</li>
-<li> Paired data analysis (case-control studies, repeated measures) </li>
+<li> Paired data analysis (matched case-control studies, repeated measures) </li>
 <li> Univariate analysis and models for continuous and categorical outcome: Correlation, linear and logistic regression</li>
 <li> Univariate analysis and models for time-dependent outcome: Kaplan-Meier curves and cox regression </li>
 <li> Multivariate analysis and models for continuous, categorical and time-dependent outcomes </li></ul>
