### 2020-10-20 #2630 @VincentAlcazer
Dear @mikldk @nistara @adithirgis ,

Thank you again for your messages and suggestions.

The paper.md has been updated taking into account these considerations:

Addition in the summary: "Other free software exist such as iNZight or Jamovi. However, while providing solutions with multiple features such as variable recoding, these software do not guide the user through the analysis and can lack some key features such as time-dependent outcome analysis."

In the Statement of need: "Compared to other free similar software, StatAid has been designed to quickly produce publication-ready graph and table by really guiding the user through their data analysis and providing multiple graph customization options. By limiting the number of choices and integrating different check and variable controls, StatAid hence help the user to prevent bad test use or bad graph choice. Besides, as an evolving software, only StatAid is providing the possibility for users to ask for the implementation of a particular feature or to contribute to the software development. Its open-source aspect can also be seen as a security for people working with sensitive data (e.g. data from clinical trials / patients). The online version of StatAid renders it accessible everywhere, even on computers with restrictive software installation policies such as hospitals or research centers."

Please let me know if you have anything else in mind to improve the paper/software.

Best regards,

Vincent
