### 2020-10-21 #2630 @adithirgis
I think this is summarised well.
I had a small suggestion if everyone else agrees (and also if I am not wrong :)) -

In the Statement of need: "Compared to other free similar software, StatAid has been designed to quickly produce publication-ready graph and table by ~really~ guiding the user through their data analysis and providing multiple graph customization options. By limiting the number of choices and integrating different check and variable controls, StatAid ~hence~ _helps_ the user to prevent bad test use or bad graph choice. Besides, as an evolving software, ~only~ StatAid ~is~ _also provides_ the possibility for users to ~ask~ _request_ for the implementation of a particular feature or to contribute to the software development. Its open-source aspect can also be seen as a security for people working with sensitive data (e.g. data from clinical trials / patients). _The online version of StatAid renders it accessible everywhere, even on computers with restrictive software installation policies such as hospitals or research centers._ (Can this sentence be modified a little?)"
