+++ paper.md
@@ -50,7 +50,7 @@ Its user-friendly interface can guide any clinician or researcher even with no p
 
 `StatAid` is an R package designed to fit the needs for every-day statistical analysis in science. The package provides all the tools necessary to perform a data analysis in an intuitive ready-to-use graphical interface. Any user with no coding skill or no software with paid-license available can easily perform all the steps of a good statistical analysis, from data-exploration/quality check to multivariate modeling. 
 
-Compared to other free similar software, `StatAid` has been designed to quickly produce publication-ready graph and table by really guiding the user through their data analysis and providing multiple graph customization options. By limiting the number of choices and integrating different check and variable controls, `StatAid` hence help the user to prevent bad test use or bad graph choice. Besides, as an evolving software, only `StatAid` is providing the possibility for users to ask for the implementation of a particular feature or to contribute to the software development. Its open-source aspect can also be seen as a security for people working with sensitive data (e.g. data from clinical trials / patients). The online version of `StatAid` renders it accessible everywhere, even on computers with restrictive software installation policies such as hospitals or research centers.
+Compared to other free similar software, `StatAid` has been designed to quickly produce publication-ready graph and table by guiding the user through their data analysis and providing multiple graph customization options. By limiting the number of choices and integrating different check and variable controls, `StatAid` helps the user to prevent bad test use or bad graph choice. Besides, as an evolving software, `StatAid` also provides the possibility for users to request for the implementation of a particular feature or to contribute to the software development. Its open-source aspect can also be seen as a security for people working with sensitive data (e.g. data from clinical trials / patients). The online version of `StatAid` makes it accessible everywhere, even on computers with restrictive policies of software installation.
 
 `StatAid` was designed to be used by clinicians, researchers, students and any person wanting to perform statistical analysis with no coding skills. It was primarily designed for medical/life science data analysis but can easily be extended to other fields.
 
