### 2020-10-22 #2630 @VincentAlcazer
Dear @mikldk @nistara @adithirgis ,

The paper.md has been updated with @adithirgis 's suggestions.

Best regards,

Vincent
