### 2020-10-26 #2630 @mikldk
@nistara, @adithirgis: Can you confirm that you have finished the review and recommend that this paper is now published?

@VincentAlcazer:

* In the paper you have "Compared to other free similar software, StatAid has been designed to quickly produce publication-ready graph and table by guiding the user through their data analysis and providing multiple graph customization options."
   + To me it sounds more correct to write "publication-ready graph**s** and table**s**", i.e. in plural, but I am not a native English speaker. Anyone that knows the correct form?
* Please make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
