### 2020-10-26 #2630 @adithirgis
Hi all!

I confirm that I have reviewed the paper and I recommend to publish this. All the best @VincentAlcazer! StatAid is an amazing and useful software.

Though I am not a good English language speaker/ user, but I do feel that the first point of plural usage seems correct.

Thanks so much again for inviting me!

Regards and Stay Safe
Adithi
