### 2020-10-27 #2630 @nistara
Hi @mikldk, @VincentAlcazer, and @adithirgis

I had previously submitted a pull request for typos and minor issues of the like in the paper/software (https://github.com/VincentAlcazer/StatAid/commit/d0afe60a93836b246158fe15b35f81567d1b8fc8), however, I am not sure if I can make extensive grammatical edits as a reviewer. I do realize/respect that as multilingual people we might require some assistance in one or the other language, and so if it's acceptable, I conducted some extra proofreading and can create another pull request for it: https://github.com/nistara/StatAid/commit/85fe82e6da936b829988c61a747db3865766c115

If it's fine by you, I would recommend this paper once the above edits have been incorporated.

Regards,
Nistara
