### 2020-10-27 #2630 @VincentAlcazer
Dear @nistara ,

Thank you for this complete proofread of the paper. If it's fine for everybody, I am waiting your pull request to integrate it and make the new Zenodo archive.

Best regards,

Vincent
