### 2020-10-28 bdc5879
minor updates to paper
