### 2020-10-28 16293ef
minor updates to paper II
