### 2020-10-28 #2630 @VincentAlcazer
Dear @mikldk ,

The paper has been completely reviewed and the software archive updated on zenodo (StatAid v1.1.0 - DOI 10.5281/zenodo.4146739).

I would like to thank again the reviewers @nistara and @adithirgis for their amazing work and the time they have taken to help me improving both the software and the paper.

Best regards,

Vincent Alcazer
