### 2020-10-28 #2630 @mikldk
@VincentAlcazer There seem to be an issue with the archive. The file available for download is named 1.1 (not 1.1.0), and at the repo there is no 1.1.0 release (<https://github.com/VincentAlcazer/StatAid/releases>).
