### 2020-10-28 #2630 @VincentAlcazer
The version has been corrected on github. The archive available on Zenodo is also the good version with the updated paper, only the name has not been changed (because the software itself was not changed).

Would you prefer that I upload a new version on Zenodo with the renamed archive (1.1.0 intead of 1.1)?
