### 2020-10-28 #2630 @mikldk
> @mikldk the version has been corrected on github. The archive available on Zenodo is also the good version with the updated paper, only the name has not been changed (because the software itself was not changed).
>
> Would you prefer that I upload a new version on Zenodo with the renamed archive (1.1.0 intead of 1.1)?

At <https://joss.readthedocs.io/en/latest/submitting.html> you can read that "Upon successful completion of the review, authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as Zenodo [...]".

So yes, the version in the paper (1.1.0) must have a tagged release of same version that is also available at Zenodo. All same version. You can integrate Zenodo and Github such that Github releases automatically get a Zenodo release.
