### 2020-10-28 #2630 @mikldk
It looks like you have folders such as `.Rproj.user` in your Git. Please remove it by creating a `.gitignore` file. Check other folders, too - I am not sure about e.g. `rsconnect`.
