### 2020-10-28 #2630 @VincentAlcazer
@mikldk  it is strange as the .gitignore file already included .Rproj.user. I manually deleted the folders and added rsconnect exclusion in the .gitignore file. It seems to work correctly now.

I had to push a new version to update the Github and Zenodo releases:
version: 1.1.01
DOI: 10.5281/zenodo.4147144
