### 2020-10-28 #2630 @mikldk
@VincentAlcazer I apologise for being pedantic, but 1.1.01 is not a valid semantic version number, which I strongly suggest to stick to. See <https://semver.org/> under [§2](https://semver.org/#spec-item-2) ("A normal version number MUST take the form X.Y.Z where X, Y, and Z are non-negative integers, and MUST NOT contain leading zeroes.").

I would suggest calling it version 1.1.1 instead.
