+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# Welcome to StatAid
+# Welcome to StatAid v1.1.2
 
 <!-- badges: start -->
 
