### 2020-10-29 #2630 @VincentAlcazer
@mikldk I understand and sorry for this fastidious release. In fact I chose 1.1.01 due to Zenodo conflicts for 1.1.1.

I updated the version (and had to change the number for Zenodo):
New version: 1.1.2
10.5281/zenodo.4152933

Thank you again for your time.

Best regards,

Vincent Alcazer
